message.dto move in chat resources
