+++ backend/src/message/message.controller.ts
@@ -8,7 +8,7 @@ import {
     Param,
 } from '@nestjs/common'
 import { MessageService } from './message.service'
-import { CreateMessageDto } from './dto/create-message.dto'
+import { CreateMessageDto } from '../chat/dto/create-message.dto'
 import { ApiTags } from '@nestjs/swagger'
 
 @ApiTags('message')
