+++ backend/src/message/message.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common'
-import { CreateMessageDto } from './dto/create-message.dto'
+import { CreateMessageDto } from '../chat/dto/create-message.dto'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Message } from 'src/typeorm/message.entity'
 import { Repository } from 'typeorm'
